requirements.txt
langchain
langchain-google-genai
google-generativeai
streamlit
streamlit-chat
tiktoken
python-dotenv